requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-Login==0.6.3
Werkzeug==2.3.7
mysql-connector-python==8.1.0
bcrypt==4.0.1
Pillow
python-dotenv==1.0.0
email-validator==2.0.0